requirements.txt
livekit
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-silero
livekit-plugins-cartesia
livekit-plugins-google
livekit-plugins-assemblyai
livekit-agents[turn-detector]
python-dotenv~=1.0
mcp
assemblyai
fastapi
uvicorn[standard]

